Revert "Server -> add some events and fix main funtions (#5)"
This reverts commit f849d75c005473fc14410a28ed0bff7e8362b982.
