+++ server/src/schemas/users.ts
@@ -5,15 +5,12 @@ const userSchema = new Schema<User>({
   phone: { type: String, required: true, unique: true },
   password: { type: String, required: true },
   username: { type: String },
-  conversationsId: { type: [], default: [] },
   options: {
     darkMode: { type: Boolean, default: false },
     online: { type: Boolean, default: true },
     lastSeen: { type: Date, default: Date.now }
   },
-  joinedAt: { type: Date, default: Date.now },
-  publicKey: { type: String },
-  socketId: { type: String }
+  joinedAt: { type: Date, default: Date.now }
 });
 
 const UserModel = model<User>("user", userSchema);
